requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
base58==2.1.1
ccxt==4.5.11
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dateparser==1.2.2
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.0
numpy==2.3.4
pandas==2.3.3
propcache==0.4.1
psutil==7.1.1
psycopg2-binary==2.9.11
pycares==4.11.0
pycparser==2.23
pycryptodome==3.23.0
pyTelegramBotAPI==4.29.1
python-binance==1.0.30
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
regex==2025.9.18
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
